+++ lib/generators/passkeys_rails/templates/passkeys_rails_config.rb
@@ -60,4 +60,47 @@ PasskeysRails.config do |c|
   # c.subscribe(:did_register) do |event, agent, request|
   #   puts("#{event} | #{agent.id} | #{request.headers}")
   # end
+
+
+  # PasskeysRails uses webauthn to help with the protocol.
+  # The following settings are passed throught webauthn.
+  # wa_origin is the only one requried
+
+
+  # This value needs to match `window.location.origin` evaluated by
+  # the User Agent during registration and authentication ceremonies.
+  # c.wa_origin = ENV['DEFAULT_HOST'] || https://myapp.mydomain.com
+
+  # Relying Party name for display purposes
+  # c.wa_relying_party_name = "My App Name"
+
+  # Optionally configure a client timeout hint, in milliseconds.
+  # This hint specifies how long the browser should wait for any
+  # interaction with the user.
+  # This hint may be overridden by the browser.
+  # https://www.w3.org/TR/webauthn/#dom-publickeycredentialcreationoptions-timeout
+  # c.wa_credential_options_timeout = 120_000
+
+  # You can optionally specify a different Relying Party ID
+  # (https://www.w3.org/TR/webauthn/#relying-party-identifier)
+  # if it differs from the default one.
+  #
+  # In this case the default would be "auth.example.com", but you can set it to
+  # the suffix "example.com"
+  #
+  # c.wa_rp_id = "example.com"
+
+  # Configure preferred binary-to-text encoding scheme. This should match the encoding scheme
+  # used in your client-side (user agent) code before sending the credential to the server.
+  # Supported values: `:base64url` (default), `:base64` or `false` to disable all encoding.
+  #
+  # c.wa_encoding = :base64url
+
+  # Possible values: "ES256", "ES384", "ES512", "PS256", "PS384", "PS512", "RS256", "RS384", "RS512", "RS1"
+  # Default: ["ES256", "PS256", "RS256"]
+  #
+  # c.wa_algorithms = ["ES256", "PS256", "RS256"]
+
+  # Append an algorithm to the existing set
+  # c.wa_algorithm = "PS512"
 end